client side + server side On the admin side, I created a “Gaming Zone” page.
On this page, when I add/select a category, the page automatically shows all products of that category.

On the client side, a page is created with the same name that we set (e.g., Gaming Zone).

On this client-side page, the products are displayed exactly like the Shop page (same layout and design).